% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_functions.R
\name{all_random}
\alias{all_random}
\title{Extract n number of baits randomly from the whole chromosome length}
\usage{
all_random(length, n, size, seed = NULL)
}
\arguments{
\item{length}{The length of the chromosome being analysed}

\item{n}{number of baits to generate}

\item{size}{The size of each bait}

\item{seed}{A number to fix the randomization process, for reproducibility}
}
\value{
a dataframe with starting and ending positions
}
\description{
Extract n number of baits randomly from the whole chromosome length
}
\keyword{internal}
