% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{assign_n_per_seq}
\alias{assign_n_per_seq}
\title{Distribute the overall number of baits by the multiple sequences}
\usage{
assign_n_per_seq(the.lengths, n, min.per.seq, size)
}
\arguments{
\item{the.lengths}{a dataframe with the length of each sequence, and the respective proportion of the overall size.}

\item{n}{Number of baits to generate (distributed across the various sequences).}

\item{min.per.seq}{Minimum number of baits per sequence. Defaults to 1.}

\item{size}{The size of each bait}
}
\value{
An updated the.lengths dataframe, with the number of baits to be extracted per sequence.
}
\description{
Distribute the overall number of baits by the multiple sequences
}
\keyword{internal}
