% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_baits.R
\name{blast_baits}
\alias{blast_baits}
\title{Blast baits against a database}
\usage{
blast_baits(x, db = NULL, blastn_args = "")
}
\arguments{
\item{x}{the output of the do_baits function}

\item{db}{OPTIONAL: A FASTA file against which the baits should be blased. If left empty, the database from which
the baits were created is used.}

\item{blastn_args}{Additional parameters that can be passed to the blastn call.}
}
\value{
The same output as do_baits, but with an additional column in the baits showing the number of matches.
}
\description{
Finds the number of matches for each bait generated by the do_baits function. Uses BLAST to match the baits against
either the same database, or a different one (if using the argument db). The blastn parameters can be adjusted using
the blastn_args argument. The following parameters are already set and should not be changed: -db, -query, -out and -outfmt.
}
