% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{check_chr_boundaries}
\alias{check_chr_boundaries}
\title{Check that start and end points fall whithin chromosome boundaries}
\usage{
check_chr_boundaries(
  exclusions = NULL,
  regions = NULL,
  targets = NULL,
  the.lengths
)
}
\arguments{
\item{exclusions}{The exclusions dataframe}

\item{regions}{The regions dataframe}

\item{targets}{The targets dataframe}

\item{the.lengths}{The lengths dataframe}
}
\value{
The lengths dataframe
}
\description{
Check that start and end points fall whithin chromosome boundaries
}
\keyword{internal}
