% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{check_chr_names}
\alias{check_chr_names}
\title{Check chromosome name consistency}
\usage{
check_chr_names(exclusions = NULL, regions = NULL, targets = NULL, the.lengths)
}
\arguments{
\item{exclusions}{The exclusions dataframe}

\item{regions}{The regions dataframe}

\item{targets}{The targets dataframe}

\item{the.lengths}{The lengths dataframe}
}
\value{
The lengths dataframe
}
\description{
Check chromosome name consistency
}
\keyword{internal}
