% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_lengths}
\alias{check_lengths}
\title{Check if the lengths dataframe follows the expected standards}
\usage{
check_lengths(lengths)
}
\arguments{
\item{lengths}{the lengths dataframe}
}
\value{
an instruction to recompile lengths or not.
}
\description{
Check if the lengths dataframe follows the expected standards
}
\keyword{internal}
