% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_handlers.R
\name{check_python}
\alias{check_python}
\title{check if Phython is installed before attempting to run python code}
\usage{
check_python()
}
\value{
logical
}
\description{
check if Phython is installed before attempting to run python code
}
\keyword{internal}
