% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_functions.R
\name{check_ranges}
\alias{check_ranges}
\title{Compare ranges' size with intended bait size}
\usage{
check_ranges(ranges, size, chr)
}
\arguments{
\item{ranges}{a dataframe with starting and ending positions}

\item{size}{The size of each bait}

\item{chr}{The chromosome name.}
}
\value{
A table with valid ranges
}
\description{
Compare ranges' size with intended bait size
}
\keyword{internal}
