% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_restrict}
\alias{check_restrict}
\title{Check if restrict is compatible with the number of sequences present}
\usage{
check_restrict(restrict, sequences)
}
\arguments{
\item{restrict}{A vector of chromosome names OR position numbers to which the analysis should be restricted.}

\item{sequences}{A vector with the names of the sequences in the current data.}
}
\value{
an updated restrict argument (numeric)
}
\description{
Check if restrict is compatible with the number of sequences present
}
\keyword{internal}
