% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{convert_line_endings}
\alias{convert_line_endings}
\title{Convert line endings from DOS to UNIX format}
\usage{
convert_line_endings(input, output = input)
}
\arguments{
\item{input}{The name of the input file}

\item{output}{The name of the output file (the same as the input by default)}
}
\value{
No return value. Called for side effects.
}
\description{
Database files must have UNIX line endings for supeRbaits to work.
}
