% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{coverage}
\alias{coverage}
\title{Compile tables with coverage values per chromosome}
\usage{
coverage(chr.lengths, baits, exclusions = NULL, combined = TRUE)
}
\arguments{
\item{chr.lengths}{The lengths of the chromosomes}

\item{baits}{A list of baits per chromosome}

\item{exclusions}{The exclusions table}

\item{combine}{Logical: Should results be displayed per type of bait?}
}
\description{
Compile tables with coverage values per chromosome
}
\keyword{internal}
