% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_functions.R
\name{expand_ranges}
\alias{expand_ranges}
\title{Transform the ranges dataframe into a sequence of numbers from which to sample}
\usage{
expand_ranges(ranges, size)
}
\arguments{
\item{ranges}{a dataframe with starting and ending positions}

\item{size}{The size of each bait}
}
\value{
A vector of valid chromosome positions
}
\description{
Transform the ranges dataframe into a sequence of numbers from which to sample
}
\keyword{internal}
