% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_nucleotides}
\alias{extract_nucleotides}
\title{Extract specific nucleotide sequences}
\usage{
extract_nucleotides(database, seq, from, to)
}
\arguments{
\item{database}{A database of sequences}

\item{seq}{The name of the sequence}

\item{from}{the first nucleotide position to be extracted}

\item{to}{the last nucleotide position to be extracted}
}
\value{
the nucleotide sequence
}
\description{
Extract specific nucleotide sequences
}
