% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_baits.R
\name{find_executable}
\alias{find_executable}
\title{Adapted from rBLAST's .findExecutable function}
\usage{
find_executable(exe)
}
\arguments{
\item{exe}{a command to be found in the system}
}
\value{
the full path to the command
}
\description{
Adapted from rBLAST's .findExecutable function
}
\keyword{internal}
