% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_functions.R
\name{find_random_ranges}
\alias{find_random_ranges}
\title{Make a table of valid ranges within the chromosome}
\usage{
find_random_ranges(starting.point, length, exclusions)
}
\arguments{
\item{starting.point}{the first valid point in the chromosome}

\item{length}{the last valid point in the chromosome}

\item{exclusions}{A subset of the exclusions dataframe for the target chromosome.}
}
\value{
A table with valid ranges
}
\description{
Make a table of valid ranges within the chromosome
}
\keyword{internal}
