% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_functions.R
\name{find_target_ranges}
\alias{find_target_ranges}
\title{Make a table of valid ranges within the chromosome}
\usage{
find_target_ranges(targets, size, length)
}
\arguments{
\item{targets}{A subset of the targets dataframe for the target chromosome}

\item{size}{The size of each bait}

\item{length}{The length of the chromosome being analysed}
}
\value{
A table with valid ranges
}
\description{
Make a table of valid ranges within the chromosome
}
\keyword{internal}
