% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{gc_table}
\alias{gc_table}
\title{Compile tables with min, mean and max GC contents per group of baits}
\usage{
gc_table(baits, combine = FALSE)
}
\arguments{
\item{baits}{A list of baits per chromosome}

\item{combine}{Logical: Should results be displayed per chromosome?}
}
\description{
Compile tables with min, mean and max GC contents per group of baits
}
\keyword{internal}
