% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_functions.R
\name{get_bait_positions}
\alias{get_bait_positions}
\title{Extract the bait start and ending points from the valid ranges}
\usage{
get_bait_positions(ranges, size, n)
}
\arguments{
\item{ranges}{a dataframe with starting and ending positions}

\item{size}{The size of each bait}

\item{n}{a vector with the same length as the number of rows in ranges, containing the number of baits to extract from each range.}
}
\value{
a dataframe with starting and ending positions
}
\description{
Extract the bait start and ending points from the valid ranges
}
\keyword{internal}
