% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_handlers.R
\name{get_lengths}
\alias{get_lengths}
\title{Python handler for lengthChrom}
\usage{
get_lengths(database, restrict = NULL)
}
\arguments{
\item{database}{A database of chromosomes}

\item{restrict}{A vector of chromosome names to which the analysis should be restricted.}
}
\value{
a data frame with the chromosome names and their lengths
}
\description{
Finds the length of each chromosome
}
\keyword{internal}
