% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_exclusions}
\alias{load_exclusions}
\title{Load file with areas of the chromosome to avoid}
\usage{
load_exclusions(file)
}
\arguments{
\item{file}{The name of the file containing the information to upload}
}
\value{
The exclusions dataframe
}
\description{
Load file with areas of the chromosome to avoid
}
\keyword{internal}
