% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_lengths}
\alias{load_lengths}
\title{Compile sequence lengths from a FASTA format file}
\usage{
load_lengths(file)
}
\arguments{
\item{file}{The name of the file containing the information to import}
}
\value{
A database containing the lengths of each sequence.
}
\description{
Compile sequence lengths from a FASTA format file
}
