% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_regions}
\alias{load_regions}
\title{Load file with areas of the chromosome to target}
\usage{
load_regions(file)
}
\arguments{
\item{file}{The name of the file containing the information to upload}
}
\value{
The regions dataframe
}
\description{
Load file with areas of the chromosome to target
}
\keyword{internal}
