% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_targets}
\alias{load_targets}
\title{Load file with points in the chromosome to target}
\usage{
load_targets(file)
}
\arguments{
\item{file}{The name of the file containing the information to upload}
}
\value{
The targets dataframe
}
\description{
Load file with points in the chromosome to target
}
\keyword{internal}
