% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_baits.R
\name{make_blast_db}
\alias{make_blast_db}
\title{copied from rBLAST's makeblastdb function}
\usage{
make_blast_db(file)
}
\arguments{
\item{file}{a FASTA file to be converted to a blast database}
}
\value{
Nothing. Creates the database files in the same location as the input file
}
\description{
copied from rBLAST's makeblastdb function
}
\keyword{internal}
