% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print_coverage}
\alias{print_coverage}
\title{Print graphics with coverage per chromosome}
\usage{
print_coverage(chr.lengths, baits, exclusions = NULL, targets = NULL)
}
\arguments{
\item{chr.lengths}{The lengths of the chromosomes}

\item{baits}{A list of baits per chromosome}

\item{exclusions}{The exclusions table}

\item{targets}{The targets table}
}
\description{
Print graphics with coverage per chromosome
}
\keyword{internal}
