% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_handlers.R
\name{retrieve_baits}
\alias{retrieve_baits}
\title{Return bps}
\usage{
retrieve_baits(chr, database)
}
\arguments{
\item{chr}{The chromosome name.}

\item{database}{A database of chromosomes}
}
\value{
logical
}
\description{
Return bps
}
