% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{standardize_lines}
\alias{standardize_lines}
\title{Ensure the database nucleotide lines have the same length}
\usage{
standardize_lines(input, output = input)
}
\arguments{
\item{input}{The name of the input file}

\item{output}{The name of the output file (the same as the input by default)}
}
\value{
No return value. Called for side effects.
}
\description{
Database files must have the same number of nucleotides per line (except for the terminal lines of each sequence).
}
