% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{subsample}
\alias{subsample}
\title{Extract the values of input that match the linking element}
\usage{
subsample(input, link)
}
\arguments{
\item{input}{A dataframe whose first column will be matched to link}

\item{link}{a keyword to be searched for}
}
\value{
A subset of input
}
\description{
Extract the values of input that match the linking element
}
\keyword{internal}
