% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superBaits-package.R
\docType{package}
\name{supeRbaits}
\alias{supeRbaits}
\alias{supeRbaits-package}
\title{supeRbaits: A package that does things}
\description{
A description of the package
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
