% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{trim_parameters}
\alias{trim_parameters}
\title{extract exclusions, regions, and targets relevant for the chromosome being analysed}
\usage{
trim_parameters(chr, exclusions = NULL, regions = NULL,
  targets = NULL)
}
\arguments{
\item{chr}{The chromosome name}

\item{exclusions}{A file containing regions to exclude}

\item{regions}{A file containing regions of interest.}

\item{targets}{A file containing bp's to target.}
}
\value{
a list of trimmed data frames
}
\description{
extract exclusions, regions, and targets relevant for the chromosome being analysed
}
\keyword{internal}
