% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_functions.R
\name{trim_ranges}
\alias{trim_ranges}
\title{Check if any range starts/ends within an exclusion zone}
\usage{
trim_ranges(ranges, exclusions)
}
\arguments{
\item{ranges}{a dataframe with starting and ending positions}

\item{exclusions}{A subset of the exclusions dataframe for the target chromosome.}
}
\value{
a dataframe with starting and ending positions
}
\description{
Check if any range starts/ends within an exclusion zone
}
\keyword{internal}
