% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_functions.R
\name{trimmed_random}
\alias{trimmed_random}
\title{Extract n number of baits randomly from parts of the chromosome length}
\usage{
trimmed_random(length, n, size, seed = NULL, exclusions)
}
\arguments{
\item{length}{The length of the chromosome being analysed}

\item{n}{number of baits to generate}

\item{size}{The size of each bait}

\item{seed}{A number to fix the randomization process, for reproducibility}

\item{exclusions}{A subset of the exclusions dataframe for the target chromosome}
}
\value{
a dataframe with starting and ending positions
}
\description{
Extract n number of baits randomly from parts of the chromosome length
}
\keyword{internal}
