% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_functions.R
\name{trimmed_targetted}
\alias{trimmed_targetted}
\title{Extract n number of baits from parts of the chromosome length, targeting specific areas}
\usage{
trimmed_targetted(length, n, size, seed = NULL, regions = NULL,
  targets = NULL, exclusions)
}
\arguments{
\item{length}{The length of the chromosome being analysed}

\item{n}{number of baits to generate}

\item{size}{The size of each bait}

\item{seed}{A number to fix the randomization process, for reproducibility}

\item{regions}{A subset of the regions dataframe for the target chromosome}

\item{targets}{A subset of the targets dataframe for the target chromosome}

\item{exclusions}{A subset of the exclusions dataframe for the target chromosome.}
}
\value{
a dataframe with starting and ending positions
}
\description{
Extract n number of baits from parts of the chromosome length, targeting specific areas
}
\keyword{internal}
