% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_functions.R
\name{check_ranges}
\alias{check_ranges}
\title{Compare ranges' size with intended bait size}
\usage{
check_ranges(ranges, n, size, chr, tiling = 1)
}
\arguments{
\item{ranges}{a dataframe with starting and ending positions}

\item{n}{number of baits to generate}

\item{size}{The size of each bait}

\item{chr}{The chromosome name, used in error messages.}

\item{tiling}{The minimum number of baits desired per range.}
}
\value{
A table with valid ranges
}
\description{
Compare ranges' size with intended bait size
}
\keyword{internal}
