% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{coverage}
\alias{coverage}
\title{Compile tables with coverage values per chromosome}
\usage{
coverage(x, combined = TRUE)
}
\arguments{
\item{x}{The output of do_baits}

\item{combine}{Logical: Should results be displayed per type of bait?}
}
\description{
Compile tables with coverage values per chromosome
}
\keyword{internal}
