% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_functions.R
\name{get_bait_positions}
\alias{get_bait_positions}
\title{Extract the bait start and ending points from the valid ranges}
\usage{
get_bait_positions(valid_ranges, seed = NULL, size, n)
}
\arguments{
\item{valid_ranges}{a sequence of valid starting points}

\item{seed}{A number to fix the randomization process, for reproducibility}

\item{size}{The size of each bait}

\item{n}{number of baits to generate}
}
\value{
a dataframe with starting and ending positions
}
\description{
Extract the bait start and ending points from the valid ranges
}
\keyword{internal}
