% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{load_lengths}
\alias{load_lengths}
\title{Load file with the chromosomes' length}
\usage{
load_lengths(file)
}
\arguments{
\item{file}{The name of the file containing the information to upload}
}
\value{
The lengths dataframe
}
\description{
Load file with the chromosomes' length
}
\keyword{internal}
