% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_functions.R
\name{main_function}
\alias{main_function}
\title{Generates random baits}
\usage{
main_function(n, size, lengths, exclusions = NULL, regions = NULL,
  targets = NULL, seed = NULL)
}
\arguments{
\item{n}{number of baits to generate}

\item{size}{The size of each bait}

\item{lengths}{A file containing the chromosomes' length}

\item{exclusions}{A file containing regions to exclude}

\item{regions}{A file containing regions to target}

\item{targets}{A file containing points to target}

\item{seed}{A number to fix the randomization process, for reproducibility}
}
\value{
A dataframe of baits
}
\description{
Function inspired by this comment in StackOverflow: https://stackoverflow.com/questions/49149839/simulate-random-positions-from-a-list-of-intervals
}
