% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{main_function}
\alias{main_function}
\title{Generates random baits}
\usage{
main_function(
  n,
  size,
  database,
  exclusions = NULL,
  regions = NULL,
  regions.prop = NULL,
  regions.tiling = NULL,
  targets = NULL,
  targets.prop = NULL,
  targets.tiling = NULL,
  seed = NULL,
  restrict = NULL,
  debug = FALSE,
  gc = c(0.3, 0.5),
  verbose = FALSE,
  useR = FALSE
)
}
\arguments{
\item{n}{number of baits to generate}

\item{size}{The size of each bait}

\item{database}{A database of chromosomes}

\item{exclusions}{A file containing regions to exclude}

\item{regions}{A file containing regions of interest.}

\item{regions.prop}{The proportion of baits that should overlap the regions of interest.}

\item{regions.tiling}{The minimum number of baits to distribute per region of interest.}

\item{targets}{A file containing bp's to target.}

\item{targets.prop}{The proportion of baits that should overlap the targets.}

\item{targets.tiling}{The minimum number of baits to distribute per target.}

\item{seed}{A number to fix the randomization process, for reproducibility}

\item{restrict}{A vector of chromosome names OR position numbers to which the analysis should be restricted.}

\item{debug}{Logical: Should debug files be created?}

\item{gc}{A vector of two values between 0 and 1, specifying the minimum and maximmum GC percentage allowed in the output baits.}

\item{verbose}{Logical: Should detailed bait processing messages be displayed per sequence?}
}
\value{
A dataframe of baits
}
\description{
Function inspired by this comment in StackOverflow: https://stackoverflow.com/questions/49149839/simulate-random-positions-from-a-list-of-intervals
}
