% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_functions.R
\name{print_coverage}
\alias{print_coverage}
\title{Print graphics with coverage per sequence}
\usage{
print_coverage(x, seq.name)
}
\arguments{
\item{x}{The output of main_function.}

\item{sequence}{the sequence to be printed.}
}
\description{
Print graphics with coverage per sequence
}
\keyword{internal}
