% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_functions.R
\name{random_baits}
\alias{random_baits}
\title{Extract n number of baits randomly from parts of the chromosome length}
\usage{
random_baits(length, n, size, seed = NULL, exclusions = NULL, chr)
}
\arguments{
\item{length}{The length of the chromosome being analysed}

\item{n}{number of baits to generate}

\item{size}{The size of each bait}

\item{seed}{A number to fix the randomization process, for reproducibility}

\item{exclusions}{A subset of the exclusions dataframe for the target chromosome.}

\item{chr}{The chromosome name.}
}
\value{
a dataframe with starting and ending positions
}
\description{
Extract n number of baits randomly from parts of the chromosome length
}
\keyword{internal}
