% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_functions.R
\name{region_baits}
\alias{region_baits}
\title{Find valid ranges within the regions}
\usage{
region_baits(chr.length, n, size, tiling = NULL, regions,
  exclusions = NULL, chr = NULL)
}
\arguments{
\item{chr.length}{The length of the chromosome being analysed}

\item{n}{number of baits to generate}

\item{size}{The size of each bait}

\item{tiling}{The minimum number of baits desired per range.}

\item{regions}{A subset of the regions dataframe for the target chromosome}

\item{exclusions}{A subset of the exclusions dataframe for the target chromosome.}

\item{chr}{The chromosome name, used in error messages.}
}
\value{
a dataframe with starting and ending positions
}
\description{
Find valid ranges within the regions
}
\keyword{internal}
