% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_handlers.R
\name{retrieve_baits}
\alias{retrieve_baits}
\title{Return bps}
\usage{
retrieve_baits(chr, positions, database)
}
\arguments{
\item{chr}{the chromosome name}

\item{positions}{the start and stop positions of the baits to be retrieved}

\item{database}{A database of chromosomes}
}
\value{
logical
}
\description{
Return bps
}
