% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superBaits-package.R
\docType{package}
\name{supeRbaits}
\alias{supeRbaits}
\alias{supeRbaits-package}
\title{supeRbaits: Sample Baits From Sequences}
\description{
Provided with a database of sequences, samples short bait sequences, 
taking into account potential exclusion or region zones, as well as specific targets.
}
\details{
Please read the package vignette for more details by running browseVignettes("supeRbaits").
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BelenJM/supeRbaits}
  \item Report bugs at \url{https://github.com/BelenJM/supeRbaits/issues}
}

}
\author{
\strong{Maintainer}: Belén Mena \email{bmen@aqua.dtu.dk} (\href{https://orcid.org/0000-0001-8458-5533}{ORCID})

Authors:
\itemize{
  \item Hugo Flávio \email{hdmfla@aqua.dtu.dk} (\href{https://orcid.org/0000-0002-5174-1197}{ORCID})
  \item Miguel Ramos \email{jmiguelsramos@gmail.com}
}

}
\keyword{internal}
